--- src/components/Modal/MODAL_TYPE.js
@@ -1,3 +1,0 @@
-export default {
-    CENTERED: 'centered',
-};
